Fixed issue of circular imports
