+++ server/app/api/v1/metadata.py
@@ -28,7 +28,7 @@ from fastapi.responses import HTMLResponse
 from server.app.db.dbconfig import get_db
 from server.app.schemas.requests import MLMDPullRequest, MLMDPushRequest
 from server.app.schemas.responses import success_response
-from server.app.main import (
+from server.app.services.mlmd_state import (
     query,
     dict_of_exe_ids,
     pipeline_locks,
